:beers: 完成了hot模块的点击跳转逻辑，修改布局

+++ module_hot/src/main/java/com/example/module_hot/viewModel/ListSongsViewModel.kt
@@ -17,7 +17,7 @@ class ListSongsViewModel: ViewModel() {
     private val _loadState = MutableStateFlow<LoadState>(LoadState.Init)
     val loadState: StateFlow<LoadState> = _loadState.asStateFlow()
 
-    fun getListSongsData(id: Int) {
+    fun getListSongsData(id: Long) {
         if (_loadState.value is LoadState.Loading){
             return
         }
